safari-534.53-branch/LayoutTests/fast/dom/Geolocation/delayed-permission-denied-for-multiple-requests-expected.txt

Tests that when multiple requests are waiting for permission, no callbacks are invoked until permission is denied.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS error.code is error.PERMISSION_DENIED
PASS error.message is "User denied Geolocation"
PASS error.code is error.PERMISSION_DENIED
PASS error.message is "User denied Geolocation"
PASS successfullyParsed is true

TEST COMPLETE
